[GROWTH BLOCK]
Changed the logic behind the crop growth.
Currently only works on wheat and such. Not cacti/sugar cane, etc.
It now work independantly from the crop. Instead of calling the updateTick of the crop more often, it replaces the top block with the same crop but different metadata.

It checks how many growth blocks below the top one (to a max of 16).
Compares against a random. 1 in (17-n) where n is the number of growth blocks.
So for a full 16 tower of growth block, the chances it will update the crops is 1 in (17-16). Or 1 in 2.

- Ace